move funções menores para utils.py
